SubDownloader Examples
These are examples (very simplified) for educational purposes, to understand and grab a template on using subtitles site's API such as podnapisi (the only one found free at the time) without requiring sign-up or creating users/passwords or even tokens.
The code is laid on top of a very simple Tkinter command to determine the path (which should contain the video's name and year).
The process simply downloads a zip file later uncompressed, the first matching subtitle file found for the name and year of the video.
The intention of this example is not to download files, but to better see how requests library works on a free API, which includes Beautiful Soup a bit, JSON formats, and some of the os library's operations, as an exercise to recall previous Python programming concepts.
This code is based on many found all over the internet as Free Code Samples. It does not copy or replicate any non-free code.